fix: compatibilidad FK notification_settings

+++ database/migrations/20260204120000_create_notification_settings.sql
@@ -2,8 +2,8 @@
 -- Timestamped: 20260204120000
 
 CREATE TABLE IF NOT EXISTS notification_settings (
-  id BIGINT AUTO_INCREMENT PRIMARY KEY,
-  user_id BIGINT NOT NULL,
+  id INT AUTO_INCREMENT PRIMARY KEY,
+  user_id INT NOT NULL,
   email_notifications BOOLEAN DEFAULT TRUE,
   weekly_digest BOOLEAN DEFAULT FALSE,
   in_app_notifications BOOLEAN DEFAULT TRUE,
